extended Kriteriumverwaltung

+++ NWAT_SS16_WPF/Controller/ControllerKriterium.cs
@@ -34,7 +34,7 @@ namespace NWAT_SS16
                 temp_objekt.setBezeichnung("Neues Kriterium");
                 db.insert(temp_objekt);
                 onUpdateView();
-                krit.listeKriterium.SelectedIndex = krit.listeKriterium.Items.Count - 1;
+                krit.listeKriterium.SelectedIndex = krit.listeKriterium.Items.Count - 1; // letztes Kriterium auswählen
             }
         }
 
@@ -44,9 +44,15 @@ namespace NWAT_SS16
             if (frm.GetType().Name == "Kriteriumverwaltung")
             {
                 Kriteriumverwaltung krit = (Kriteriumverwaltung)frm;
-                List<Kriterium> items = db.get(new Kriterium());
-                krit.listeKriterium.ItemsSource = items;
-                
+                List<Kriterium> kriterien = db.get(new Kriterium());
+                krit.listeKriterium.ItemsSource = kriterien;
+
+                List<Kriterium> oberkriterien = kriterien;
+                Kriterium temp_objekt = new Kriterium();
+                temp_objekt.setBezeichnung("Kein Oberkriterium");
+                oberkriterien.Add(temp_objekt);
+                krit.details_OberKriterium.ItemsSource = oberkriterien;
+                krit.details_OberKriterium.SelectedIndex = krit.details_OberKriterium.Items.Count - 1; // letztes OberKriterium auswählen
             }
         }
 
@@ -55,9 +61,15 @@ namespace NWAT_SS16
             if (frm.GetType().Name == "Kriteriumverwaltung")
             {
                 Kriteriumverwaltung krit = (Kriteriumverwaltung)frm;
-                List<Kriterium> items = db.get(new Kriterium());
-                krit.listeKriterium.ItemsSource = items;
+                List<Kriterium> kriterien = db.get(new Kriterium());
+                krit.listeKriterium.ItemsSource = kriterien;
 
+                List<Kriterium> oberkriterien = kriterien;
+                Kriterium temp_objekt = new Kriterium();
+                temp_objekt.setBezeichnung("Kein Oberkriterium");
+                oberkriterien.Add(temp_objekt);
+                krit.details_OberKriterium.ItemsSource = oberkriterien;
+                krit.details_OberKriterium.SelectedIndex = krit.details_OberKriterium.Items.Count - 1; // letztes OberKriterium auswählen
             }
         }
 
@@ -76,8 +88,15 @@ namespace NWAT_SS16
             {
                 Kriterium temp_objekt = (Kriterium)objekt;
                 Kriteriumverwaltung krit = (Kriteriumverwaltung)frm;
+
                 krit.details_ID.Text = temp_objekt.getKriteriumID().ToString();
                 krit.details_Bezeichnung.Text = temp_objekt.getBezeichnung();
+                krit.details_ProduktID.Text = "0"; // Standard
+                krit.details_ProjektID.Text = "0"; // Standard
+                krit.kriterium_aendern.IsEnabled = true;
+                krit.kriterium_loeschen.IsEnabled = true;
+                krit.details_OberKriterium.IsEnabled = true;
+                krit.details_Bezeichnung.IsEnabled = true;
             }
         }
 
@@ -87,10 +106,22 @@ namespace NWAT_SS16
             {
                 return;
             }
-            if (MessageBox.Show("Sind Sie sich sicher, dass sie das ausgewählte Kriterium löschen wollen?", "Löschen", MessageBoxButton.YesNo, MessageBoxImage.Warning) == MessageBoxResult.Yes)
+            if (frm.GetType().Name == "Kriteriumverwaltung")
             {
-                db.delete(objekt);
-                onUpdateView();
+                Kriteriumverwaltung krit = (Kriteriumverwaltung)frm;
+                if (MessageBox.Show("Sind Sie sich sicher, dass sie das ausgewählte Kriterium löschen wollen?", "Löschen", MessageBoxButton.YesNo, MessageBoxImage.Warning) == MessageBoxResult.Yes)
+                {
+                    db.delete(objekt);
+                    krit.kriterium_aendern.IsEnabled = false;
+                    krit.kriterium_loeschen.IsEnabled = false;
+                    krit.details_OberKriterium.IsEnabled = false;
+                    krit.details_Bezeichnung.IsEnabled = false;
+                    krit.details_Bezeichnung.Text = "";
+                    krit.details_ID.Text = "";
+                    krit.details_ProduktID.Text = "";
+                    krit.details_ProjektID.Text = "";
+                    onUpdateView();
+                }
             }
         }
